Take into consideration the elevation of the level
Each piece of furniture in SH3D has it's position in 3D space including
its elevation. That the elevation is relative to the floor of the level
it's placed in.
When calculating the 2D-projected location of an entity we observe the
building in 3D space with all of its levels so it wasn't enough to use
the relative elevation of the entity but we need to offset it with the
elevation of the entire floor.

Fixes #82, for real (I hope)