Add rounding of entity positions
The position which is stored in an entity's properties may fractionally
differ from the position set in SH3D.
E.g. `(1, 1)` in SH3D may be `(0.99, 1.01)` in the properties.

This means, although entities are in the same position in SH3D, their
positions differ in the script. This is fixed by rounding the x and y
values to their nearest integer value.