Add cluster movement
In case of an entity being a cluster, the cluster needs to move as a
whole.

Having a direction length of 0, shouldn't be happening anymore, because
entities which originally had the same position are now grouped inside a
cluster.
As a backup, we skip the statements for moving the entity;
Moving something to nowhere, isn't moving anyway.