Fix icons placement edge-changes
When two or more pieces of furniture have the same position and are
associated with an HA entity, they overlap each other perfectly and so
the algorithm that tries to move them doesn't know where to. Added a
hack that in this case, they be moved by a pseudo-random direction.